chore: cast stuff

+++ barretenberg/cpp/src/barretenberg/dsl/acir_format/ecdsa_secp256k1.test.cpp
@@ -41,34 +41,34 @@ size_t generate_ecdsa_constraint(EcdsaSecp256k1Constraint& ecdsa_constraint, Wit
     std::array<uint32_t, 32> pub_x_indices_in;
     std::array<uint32_t, 32> pub_y_indices_in;
     std::array<uint32_t, 64> signature_in;
-    size_t offset = 0;
+    uint32_t offset = 0;
     for (uint32_t i = 0; i < hashed_message.size(); ++i) {
         uint32_t[i] = (i + offset);
         const auto byte = static_cast<uint8_t>(hashed_message[i]);
         witness_values.emplace_back(byte);
     }
-    offset += message_in.size();
+    offset += static_cast<uint32_t>(message_in.size());
 
     for (uint32_t i = 0; i < 32; ++i) {
         pub_x_indices_in[i] = (i + offset);
         witness_values.emplace_back(pub_x_value.slice(248 - i * 8, 256 - i * 8));
     }
-    offset += pub_x_indices_in.size();
+    offset += static_cast<uint32_t>(pub_x_indices_in.size());
     for (uint32_t i = 0; i < 32; ++i) {
         pub_y_indices_in[i] = (i + offset);
         witness_values.emplace_back(pub_y_value.slice(248 - i * 8, 256 - i * 8));
     }
-    offset += pub_y_indices_in.size();
+    offset += static_cast<uint32_t>(pub_y_indices_in.size();
     for (uint32_t i = 0; i < 32; ++i) {
         signature_in[i] = (i + offset);
         witness_values.emplace_back(signature.r[i]);
     }
-    offset += signature.r.size();
+    offset += static_cast<uint32_t>(signature.r.size());
     for (uint32_t i = 0; i < 32; ++i) {
         signature_in[i] = (i + offset);
         witness_values.emplace_back(signature.s[i]);
     }
-    offset += signature.s.size();
+    offset += static_cast<uint32_t>(signature.s.size());
 
     witness_values.emplace_back(1);
     const auto result_in = static_cast<uint32_t>(offset);
